backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
solana==0.30.2
requests==2.31.0
pydantic==2.4.2
websockets==11.0.3
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6 